+++ config/application.rb
@@ -23,5 +23,8 @@ module Balance
 
     # For not swallow errors in after_commit/after_rollback callbacks.
     config.active_record.raise_in_transactional_callbacks = true
+
+    # Add export module to path
+    config.autoload_paths << Rails.root.join('app', 'models', 'export')
   end
 end