+++ config/initializers/app_config.rb
@@ -1,12 +1,12 @@
+# frozen_string_literal: true
+
 require 'csv'
 
-Time::DATE_FORMATS[:on_at] = "on %b %d, %Y at %I:%M %p"
-Time::DATE_FORMATS[:on] = "on %m.%d.%Y"
-Time::DATE_FORMATS[:date] = "%m.%d.%Y"
-Time::DATE_FORMATS[:long_date] = "%d %B %Y"
-Time::DATE_FORMATS[:time] = "%l:%M %p"
-
-Date::DATE_FORMATS[:long_date] = "%d %B %Y"
+Time::DATE_FORMATS[:on_at] = 'on %b %d, %Y at %I:%M %p'
+Time::DATE_FORMATS[:on] = 'on %m.%d.%Y'
+Time::DATE_FORMATS[:date] = '%m.%d.%Y'
+Time::DATE_FORMATS[:long_date] = '%d %B %Y'
+Time::DATE_FORMATS[:time] = '%l:%M %p'
 
 class Integer
   def to_currency