+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Started rewriting views using haml
   - Added a task to import data from CSVs
     - `bundle exec rake backup:import FOLDER=20160320173700`
+  - Users can backup data from their account
 - **Entry Changes**
   - Entry description fields now expand automatically to fit longer descriptions
   - Hovering mouse over entries now highlights the entry on the calendar